store maildir filepath in hcache
maildir uses h->path to store the filename of the email.
It's currently not stored in hcache and therefore needs to be saved and
restored in an extra step.
From here on path is stored in hcache. Invalidates old caches!

Additionally there's no need to save/restore h->old outside
mutt_hcache_restore(), it already gets restored properly.